# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /class-sessions:
    get:
      tags: [Bookings]
      operationId: listClassSessions
      summary: List real, per-box class sessions in a date range for Agendar
      description: |
        Replaces the mobile app's old fixed mock schedule — sessions come
        from the SAME class_sessions/class_bookings tables box-admin's own
        class-scheduling UI manages. Resolves `userId`'s box via
        box_members; an athlete with no box gets an empty list, never
        another box's (or nobody's) schedule.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
          description: ISO date (YYYY-MM-DD), inclusive.
        - name: to
          in: query
          required: true
          schema:
            type: string
          description: ISO date (YYYY-MM-DD), inclusive.
      responses:
        "200":
          description: Sessions in range for this athlete's box (empty if they have none).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSessionDto"
        "400":
          description: Missing or invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /bookings:
    get:
      tags: [Bookings]
      operationId: listBookings
      summary: List a user's active class bookings and waitlist entries
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active bookings and waitlist entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingRecord"
        "400":
          description: Missing user id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    post:
      tags: [Bookings]
      operationId: createBooking
      summary: Book a class or join its FIFO waitlist
      description: |
        `sessionId` must be a real class_sessions.id belonging to the
        caller's own box (see GET /class-sessions) — capacity and the
        confirmed count are read from the database, never trusted from the
        client. If no seat remains, the user is added to a maximum-five-
        person FIFO waitlist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "200":
          description: Booking or waitlist entry created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingActionResponse"
        "400":
          description: Invalid booking request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: The class is full and its waitlist has five people.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /bookings/cancel:
    post:
      tags: [Bookings]
      operationId: cancelBooking
      summary: Cancel a booking or leave a class waitlist
      description: |
        Cancelling a confirmed booking promotes the earliest waiting student
        atomically and creates an in-app notification for that student.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelBookingRequest"
      responses:
        "200":
          description: Booking cancelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingActionResponse"
        "400":
          description: Invalid cancellation request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Booking not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreateBookingRequest:
      type: object
      required: [sessionId, userId]
      properties:
        sessionId:
          type: string
          minLength: 1
        userId:
          type: string
          minLength: 1
    CancelBookingRequest:
      type: object
      required: [sessionId, userId]
      properties:
        sessionId:
          type: string
          minLength: 1
        userId:
          type: string
          minLength: 1
    BookingRecord:
      type: object
      required: [id, sessionId, userId, status, createdAt, position]
      properties:
        id:
          type: string
        sessionId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [confirmed, waiting]
        createdAt:
          type: string
        position:
          type: integer
          nullable: true
    BookingActionResponse:
      type: object
      required: [sessionId, userId, status, position, promotedUserId]
      properties:
        sessionId:
          type: string
        userId:
          type: string
        status:
          type: string
          enum: [confirmed, waiting, cancelled]
        position:
          type: integer
          nullable: true
        promotedUserId:
          type: string
          nullable: true
    ClassSessionDto:
      type: object
      description: |
        A real class_sessions row (the same table box-admin's class-scheduling
        UI reads/writes), enriched with this specific viewer's booking status
        for it and the confirmed roster's names.
      required:
        [
          id,
          boxId,
          name,
          date,
          startTime,
          durationMinutes,
          capacity,
          level,
          coachName,
          confirmedCount,
          remaining,
          myStatus,
          myWaitlistPosition,
          attendeeNames,
        ]
      properties:
        id:
          type: string
        boxId:
          type: string
        name:
          type: string
        date:
          type: string
          description: ISO date (YYYY-MM-DD).
        startTime:
          type: string
          description: "HH:MM, 24h."
        durationMinutes:
          type: integer
        capacity:
          type: integer
        level:
          type: string
        coachName:
          type: string
          nullable: true
        confirmedCount:
          type: integer
        remaining:
          type: integer
        myStatus:
          type: string
          enum: [none, confirmed, waiting]
        myWaitlistPosition:
          type: integer
          nullable: true
        attendeeNames:
          type: array
          items:
            type: string
...